Push source codes from desktop

+++ dev/src/pages/Playlist/playlistDashboard.js
@@ -1,27 +1,47 @@
-import { Card, Col, Container, Image, Row, Button } from 'react-bootstrap';
+import { Card, Col, Container, Row } from 'react-bootstrap';
 import { useEffect, useState } from "react";
-import { Link, useNavigate, useParams } from "react-router-dom";
-import { PencilSquare, PlusLg, Trash3Fill } from 'react-bootstrap-icons';
+import { useNavigate } from "react-router-dom";
+import { PlusLg, Trash3Fill } from 'react-bootstrap-icons';
 import "./playlistdashboard.css";
 
-function DisplayPlaylist() {
+function DisplayPlaylist({userId}) {
 
-    const [playlists, setPlaylist] = useState([]);
+    const [playlists, setPlaylists] = useState([]);
     const navigate = useNavigate();
 
-    const { id } = useParams();
+    const id  = userId || localStorage.getItem("userId");
+
+    const getAllPlaylistsData = async() => {
+        try {
+            const respsone = await fetch(`http://localhost:5000/api/playlists-data/${id}`, {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                },
+            })
+            const data = await respsone.json();
+            setPlaylists(data.playlistsWithTracks);
+
+        } catch (error) {
+            console.error("Error to get all playlists data: ", error.message);
+        }
+    }
 
     const fetchPlaylist = async () => {
         try {
-            const response = await fetch(`http://localhost:5000/api/playlist/u/${id}`);
+            const response = await fetch(`http://localhost:5000/api/playlist/u/${id}`,{
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                },
+            });
             const data = await response.json();
 
-            console.log("Fetched playlist data:", data);
             if (Array.isArray(data)) {
-                setPlaylist(data);
+                setPlaylists(data);
             } else {
                 console.warn("No playlists found or unexpected response:", data);
-                setPlaylist([]); // fallback so `.map()` doesn't crash
+                setPlaylists([]); // fallback so `.map()` doesn't crash
             }
         } catch (error) {
             console.error("error while fetching playlists:", error.message);
@@ -29,33 +49,30 @@ function DisplayPlaylist() {
     }
 
     useEffect(() => {
-        fetchPlaylist();
+        getAllPlaylistsData();
     }, []);
-
-    const handleUpdate = (playlistId) => {
-        navigate(`/u/${id}/playlist/${playlistId}`);
-    }
 
     const handleAdd = () => {
         navigate(`/u/${id}/playlist`);
     }
 
-    const handlePlaylistDetails = (playlistId) => {
-        // localStorage.setItem('playlistId', playlistId);
-        navigate(`/u/${id}/playlist/${playlistId}/playlist-details`, {state: {playlistId: playlistId}});
+    const handlePlaylistDetails = (playlist) => {
+        navigate(`/u/${id}/playlist/${playlist._id}/playlist-details`, {state: {playlist: playlist}});
     }
 
     const handleDelete = async (playlistId) => {
         try {
-            const response = await fetch(`http://localhost:5000/api/playlist/${playlistId}`, {
+            if (window.confirm("Do you want to delete this playlist?")) {
+                const response = await fetch(`http://localhost:5000/api/playlist/${playlistId}`, {
                 method: "DELETE",
                 headers: {
                     "Content-Type": "application/json",
                 }
-            });
-            if (response.ok) {
-                fetchPlaylist();
-            }
+                });
+                if (response.ok) {
+                    window.location.reload();
+                }
+            }    
         } catch (error) {
             console.error("error while deleting playlist:", error.message);
         }
@@ -66,24 +83,28 @@ function DisplayPlaylist() {
             <div className='custom-playlist-body'>
                 <Container>
                     <Row className="gx-3 gy-4">
-                        {playlists.map((playlist) => (
+                        {playlists && playlists?.map((playlist) => (
                             <Col key={playlist._id} xs={12} sm={6} md={4} lg={3}>
-                                <Card className="custom-playlist-card" onClick={() => handlePlaylistDetails(playlist._id)}>
+                                <Card className="custom-playlist-card" onClick={() => handlePlaylistDetails(playlist)}>
                                     <div className="icon-container">
-                                        <PencilSquare
-                                            className="icon edit-icon"
-                                            onClick={() => handleUpdate(playlist._id)}
-                                        />
                                         <Trash3Fill
                                             className="icon delete-icon"
+                                            size={24}
                                             onClick={() => handleDelete(playlist._id)}
                                         />
                                     </div>
 
-                                    <Card.Img
-                                        src="/sea.jpg"
-                                        className="custom-playlist-img"
-                                    />
+                                    <div className="playlist-cover-grid">
+                                            {playlist?.tracks.slice(0, 4).map((track, index) => (
+                                                <img
+                                                    key={index}
+                                                    src={track.image}
+                                                    alt={`track-${index}`}
+                                                    className="playlist-cover-img"
+                                                />
+                                            ))}
+                                    </div>
+
 
                                     <Card.Body>
                                         <Card.Title className="custom-playlist-title">
@@ -96,7 +117,7 @@ function DisplayPlaylist() {
                                             Created by: {playlist.userId}
                                         </Card.Text> */}
                                         <Card.Text className="custom-playlist-date">
-                                            {playlist.created_date}
+                                            {playlist?.tracks?.length} Songs
                                         </Card.Text>
                                     </Card.Body>
                                 </Card>
@@ -107,7 +128,7 @@ function DisplayPlaylist() {
                         <Col xs={12} sm={6} md={4} lg={3}>
                             <Card
                                 className="add-card"
-                                onClick={handleAdd}
+                                onClick={() =>handleAdd()}
                             >
                                 <div className="add-content">
                                     <PlusLg className="add-icon" />
